Updated PostSharp; Extacted the Logger Aspect into a separate project
