+++ PostSharpTutorial/ConsoleExample/ConsoleLogger.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using LoggerAspect;
 
 namespace ConsoleExample
 {
