+++ PostSharpTutorial/ConsoleExample/packages.config
@@ -1,4 +1,4 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="PostSharp" version="4.2.18" targetFramework="net46" />
+  <package id="PostSharp" version="4.2.20" targetFramework="net46" />
 </packages>
