+++ PostSharpTutorial/PostSharpTutorial.sln
@@ -5,6 +5,8 @@ VisualStudioVersion = 14.0.24720.0
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ConsoleExample", "ConsoleExample\ConsoleExample.csproj", "{392D441C-306A-4861-AEB9-94A68F8E58DC}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LoggerAspect", "LoggerAspect\LoggerAspect.csproj", "{8FF41B7E-0E37-4341-9020-07EBE8B43DA8}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -15,6 +17,10 @@ Global
 		{392D441C-306A-4861-AEB9-94A68F8E58DC}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{392D441C-306A-4861-AEB9-94A68F8E58DC}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{392D441C-306A-4861-AEB9-94A68F8E58DC}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8FF41B7E-0E37-4341-9020-07EBE8B43DA8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8FF41B7E-0E37-4341-9020-07EBE8B43DA8}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8FF41B7E-0E37-4341-9020-07EBE8B43DA8}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8FF41B7E-0E37-4341-9020-07EBE8B43DA8}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
